# sidecar/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "edgar-sidecar",
    "summary": "HTTP wrapper around edgartools - the single SEC egress",
    "version": "0.1.0"
  },
  "paths": {
...
    "/company/{id}/financials": {
      "get": {
        "summary": "Get Company Financials",
        "operationId": "get_company_financials_company__id__financials_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "annual",
                "quarterly"
              ],
              "type": "string",
              "default": "annual",
              "title": "Period"
            }
          },
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "raw",
                "standardized"
              ],
              "type": "string",
              "default": "standardized",
              "title": "View"
            }
          },
          {
            "name": "dimensions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dimensions"
            }
          },
          {
            "name": "amendments",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Amendments"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialsResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
...
    "/company/{id}/financials/multi": {
      "get": {
        "summary": "Get Company Financials Multi",
        "operationId": "get_company_financials_multi_company__id__financials_multi_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "annual",
                "quarterly"
              ],
              "type": "string",
              "default": "annual",
              "title": "Period"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 8,
              "minimum": 2,
              "default": 4,
              "title": "N"
            }
          },
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "raw",
                "standardized"
              ],
              "type": "string",
              "default": "standardized",
              "title": "View"
            }
          },
          {
            "name": "dimensions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dimensions"
            }
          },
          {
            "name": "amendments",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Amendments"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiFinancialsResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
    "/company/{id}/financials/ttm": {
      "get": {
        "summary": "Get Company Financials Ttm",
        "operationId": "get_company_financials_ttm_company__id__financials_ttm_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "concept",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Concept"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TTMResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Canonical body for every non-2xx the sidecar emits (HTTPException, mapped\nedgartools/httpx exceptions, and flattened request-validation errors alike)."
      },
...
      "FilingProvenance": {
        "properties": {
          "form": {
            "type": "string",
            "title": "Form"
          },
          "accession_number": {
            "type": "string",
            "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
            "title": "Accession Number"
          },
          "filing_date": {
            "type": "string",
            "format": "date",
            "title": "Filing Date"
          },
          "period_of_report": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Of Report"
          },
          "superseded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By",
            "description": "Accession of the latest later-filed amendment covering the same report period; null when this filing stands as-is."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "form",
          "accession_number",
          "filing_date",
          "period_of_report",
          "superseded_by"
        ],
        "title": "FilingProvenance"
      },
...
      "FinancialStatement": {
        "properties": {
          "periods": {
            "items": {
              "$ref": "#/components/schemas/StatementPeriod"
            },
            "type": "array",
            "title": "Periods"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/StatementRecord"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "periods",
          "records"
        ],
        "title": "FinancialStatement"
      },
      "FinancialsResponse": {
        "properties": {
          "cik": {
            "type": "string",
            "pattern": "^\\d{10}$",
            "title": "Cik"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "form": {
            "type": "string",
            "title": "Form"
          },
          "accession_number": {
            "type": "string",
            "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
            "title": "Accession Number"
          },
          "filing_date": {
            "type": "string",
            "format": "date",
            "title": "Filing Date"
          },
          "period_of_report": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Of Report"
          },
          "superseded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By",
            "description": "Accession of the latest later-filed amendment covering the same report period; null when this filing stands as-is."
          },
          "period": {
            "type": "string",
            "enum": [
              "annual",
              "quarterly"
            ],
            "title": "Period"
          },
          "view": {
            "type": "string",
            "enum": [
              "raw",
              "standardized"
            ],
            "title": "View"
          },
          "dimensions": {
            "type": "boolean",
            "title": "Dimensions"
          },
          "amendments": {
            "type": "boolean",
            "title": "Amendments"
          },
          "income_statement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          },
          "balance_sheet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          },
          "cashflow_statement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          },
          "statement_of_equity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          },
          "comprehensive_income": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          },
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cik",
          "company",
          "form",
          "accession_number",
          "filing_date",
          "period_of_report",
          "superseded_by",
          "period",
          "view",
          "dimensions",
          "amendments",
          "income_statement",
          "balance_sheet",
          "cashflow_statement",
          "statement_of_equity",
          "comprehensive_income",
          "cover"
        ],
        "title": "FinancialsResponse"
      },
...
      "MultiFinancialsResponse": {
        "properties": {
          "cik": {
            "type": "string",
            "pattern": "^\\d{10}$",
            "title": "Cik"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "period": {
            "type": "string",
            "enum": [
              "annual",
              "quarterly"
            ],
            "title": "Period"
          },
          "view": {
            "type": "string",
            "enum": [
              "raw",
              "standardized"
            ],
            "title": "View"
          },
          "dimensions": {
            "type": "boolean",
            "title": "Dimensions"
          },
          "amendments": {
            "type": "boolean",
            "title": "Amendments"
          },
          "filings": {
            "items": {
              "$ref": "#/components/schemas/FilingProvenance"
            },
            "type": "array",
            "title": "Filings"
          },
          "income_statement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StitchedFinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          },
          "balance_sheet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StitchedFinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          },
          "cashflow_statement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StitchedFinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          },
          "statement_of_equity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StitchedFinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          },
          "comprehensive_income": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StitchedFinancialStatement"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cik",
          "company",
          "period",
          "view",
          "dimensions",
          "amendments",
          "filings",
          "income_statement",
          "balance_sheet",
          "cashflow_statement",
          "statement_of_equity",
          "comprehensive_income"
        ],
        "title": "MultiFinancialsResponse"
      },
...
      "StatementPeriod": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "period_type": {
            "type": "string",
            "enum": [
              "duration",
              "instant"
            ],
            "title": "Period Type"
          },
          "period_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "period_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Months",
            "description": "Approximate duration in whole months (3=quarter, 12=annual); null for instant periods."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "label",
          "period_type",
          "period_start",
          "period_end",
          "period_months"
        ],
        "title": "StatementPeriod"
      },
...
      "StitchedFinancialStatement": {
        "properties": {
          "periods": {
            "items": {
              "$ref": "#/components/schemas/StatementPeriod"
            },
            "type": "array",
            "title": "Periods"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/StitchedStatementRecord"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "periods",
          "records"
        ],
        "title": "StitchedFinancialStatement"
      },
...
      "TTMMetricModel": {
        "properties": {
          "concept": {
            "type": "string",
            "title": "Concept"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "public_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Date",
            "description": "Latest filing_date across the facts used in the calculation: when this data vintage was fully on file. The library may back a historical window with comparative facts re-reported in later filings, so this is the vintage's publication date, not necessarily the earliest date a TTM for the window was knowable. Null when fact provenance is missing."
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/TTMPeriod"
            },
            "type": "array",
            "title": "Periods"
          },
          "has_gaps": {
            "type": "boolean",
            "title": "Has Gaps"
          },
          "has_calculated_q4": {
            "type": "boolean",
            "title": "Has Calculated Q4"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "concept",
          "label",
          "value",
          "unit",
          "as_of_date",
          "public_date",
          "periods",
          "has_gaps",
          "has_calculated_q4",
          "warning"
        ],
        "title": "TTMMetricModel"
      },
      "TTMPeriod": {
        "properties": {
          "fiscal_year": {
            "type": "integer",
            "title": "Fiscal Year"
          },
          "fiscal_period": {
            "type": "string",
            "title": "Fiscal Period"
          },
          "filing_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filing Date"
          },
          "accession_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accession Number"
          },
          "form_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fiscal_year",
          "fiscal_period",
          "filing_date",
          "accession_number",
          "form_type"
        ],
        "title": "TTMPeriod"
      },
      "TTMResponse": {
        "properties": {
          "cik": {
            "type": "string",
            "pattern": "^\\d{10}$",
            "title": "Cik"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "concept": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept"
          },
          "revenue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TTMMetricModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "net_income": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TTMMetricModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "metric": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TTMMetricModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cik",
          "company",
          "as_of",
          "concept",
          "revenue",
          "net_income",
          "metric"
        ],
        "title": "TTMResponse"
      },
...
    }
  }
}